Updated rating model to use decimal not float

+++ movie_rating/ratings/models.py
@@ -117,6 +117,6 @@ class Rating(models.Model):
         related_name='ratings',
         on_delete=models.CASCADE
     )
-    rating = models.FloatField()
+    rating = models.DecimalField(max_digits=2, decimal_places=2)
     comment = models.TextField(null=True, blank=True)
     timestamp = models.DateTimeField(auto_now_add=True)
